Refactor MainActivity setup and tuning helpers into separate file
